Store all annotations in Kotlin metadata on JVM
Annotations are declared in the common `metadata.proto`, but at the
moment, are only serialized on JVM. There are proto extensions for
annotations in klib, but hopefully they will be migrated to annotations
in the common proto in KT-64237.

At the moment, the stored annotations are not used directly anywhere in
the compiler. The only place where they're used is protoDifferenceUtils,
which makes incremental compilation correctly track difference in
annotations, which is what the changes in IC test data are about. The
annotations in metadata will also be used in a subsequent commit in
kotlinx-metadata-jvm, as well as tests on kotlinp and JVM ABI
differences.

The new fields are written since metadata version 2.2. Note that the new
fields are not read yet until KT-75020, so this does not hinder our
one-version forward compatibility, by which Kotlin 2.1 can read binaries
of metadata version 2.2 in the same way as Kotlin 2.2 can.

 #KT-57919 Fixed