FE: Prevent from visiting arguments of `a[i, ..] += e` twice
This construction is desugared into a block where all arguments
are being desugared into variables which might be used both for
`get` and `set` call.

But the problem is that at the stage where variables are created,
relevant initializer expressions have already been visited during
regular argument analysis.

Imagine we've got `a[i, .., {}] += e`
At first, we would resolve as the `i, .., {}` as arguments
of the `get` call.

After that, we desugar all the _index_ things as separate variables:
- val a0 = i
..
- val aN = {}

And then we transform them via regular body transformers to change
their status and property type properly. So, this is where the problem
appears: we start visiting the same expression for the second time.

Previously, it wasn't critical, but after the re-work for KT-67869,
some contracts are violated by this duplicated resolution.